gesperrt_car layer may be toggled now

+++ plugins/SRTShortcuts.pm
@@ -25,12 +25,12 @@ BEGIN {
 
 use strict;
 use vars qw($VERSION);
-$VERSION = 2.11;
+$VERSION = 2.12;
 
 use File::Glob qw(bsd_glob);
 
 use your qw(%MultiMap::images $BBBikeLazy::mode
-	    %main::line_width %main::p_width %main::str_draw %main::p_draw
+	    %main::line_width %main::p_width %main::str_draw %main::p_draw %main::p_file
 	    %main::p_obj
 	    $main::lazy_plot %main::lazy_p %main::layer_active_color
 	    %main::add_net
@@ -430,11 +430,10 @@ EOF
 				  maybe_orig_file => 0, # does not exist in non-orig version, already resoved to strassen+handicap
 				  above => $str_layer_level,
 				 ),
-		[Button => $do_compound->("gesperrt_car", $images{car_cross}), -command => sub { add_new_nonlazy_maybe_orig_layer("sperre", "gesperrt_car", -interactivelyselected => 0) }],
-## XXX no support for "sperre" type yet:
-#		layer_checkbutton([$do_compound->('gesperrt_car', $images{car_cross})]
-#				  'sperre', 'gesperrt_car,
-#				  maybe_orig_file => 1),
+		layer_checkbutton([$do_compound->('gesperrt_car', $images{car_cross})],
+				  'sperre', 'gesperrt_car',
+				  maybe_orig_file => 1,
+				  above => $str_layer_level),
 		layer_checkbutton([$do_compound->('Gerichtete Handicaps')],
 				  'str', "$main::datadir/handicap_directed",
 				  maybe_orig_file => 1,
@@ -629,7 +628,10 @@ EOF
 				    "$bbbike_auxdir/vmz/bbd/strassen",
 				    below => $str_layer_level,
 				   ),
-		  [Button => 'gesperrt', -command => sub { add_new_nonlazy_layer('sperre', "$bbbike_auxdir/vmz/bbd/gesperrt") }],
+		  layer_checkbutton('gesperrt', 'sperre',
+				    "$bbbike_auxdir/vmz/bbd/gesperrt",
+				    above => $str_layer_level,
+				   ),
 		  layer_checkbutton('qualitaet', 'str',
 				    "$bbbike_auxdir/vmz/bbd/qualitaet_s",
 				    above => $str_layer_level,
@@ -1132,15 +1134,21 @@ sub add_new_layer {
 	}
     }
     $layer_for_type_file{"$type $file"} = $free_layer;
-    if (!$BBBikeLazy::mode) {
-	require BBBikeLazy;
-	BBBikeLazy::bbbikelazy_empty_setup();
-	main::handle_global_directives($file, $free_layer);
-	main::bbbikelazy_add_data($type, $free_layer, $file);
-	main::bbbikelazy_init();
+    if ($type eq 'sperre') { # XXX BBBikeLazy does not handle sperre type yet
+	$main::p_draw{$free_layer} = 1;
+	$main::p_file{$free_layer} = $file; # XXX not needed for plot_sperre, but for the LayerEditor (to display the file name as label)
+	main::plot_sperre($file, -abk => $free_layer);
     } else {
-	main::handle_global_directives($file, $free_layer);
-	main::bbbikelazy_add_data($type, $free_layer, $file);
+	if (!$BBBikeLazy::mode) {
+	    require BBBikeLazy;
+	    BBBikeLazy::bbbikelazy_empty_setup();
+	    main::handle_global_directives($file, $free_layer);
+	    main::bbbikelazy_add_data($type, $free_layer, $file);
+	    main::bbbikelazy_init();
+	} else {
+	    main::handle_global_directives($file, $free_layer);
+	    main::bbbikelazy_add_data($type, $free_layer, $file);
+	}
     }
     # XXX add_to_stack functionality gots destroyed by calling once the layer_editor, because it used special_raise for *all*. get rid of special_raise/lower!
     main::add_to_stack($free_layer, "before", "pp");
@@ -1170,23 +1178,29 @@ sub toggle_new_layer {
 	    main::status_message("Cannot toggle layer: $@", 'die');
 	}
     } else {
-	eval {
-	    $layer = $layer_for_type_file{$type_file};
-	    delete $main::str_draw{$layer};
+	$layer = $layer_for_type_file{$type_file};
+	if ($type eq 'sperre') { # XXX BBBikeLazy does not handle sperre type yet
 	    delete $main::p_draw{$layer};
-	    if ($type eq 'p') {
-		$main::c->delete($layer.'-fg');
-		$main::c->delete($layer.'-img');
-	    } else {
-		$main::c->delete($layer);
-	    }
+	    main::plot_sperre(undef, -abk => $layer);
 	    delete $layer_for_type_file{$type_file};
-	    BBBikeLazy::bbbikelazy_remove_data($type, $layer);
-	    $active = 0;
-	};
-	if ($@) {
-	    $want_plot_type_file{$type_file} = 1;
-	    main::status_message("Cannot toggle layer: $@", 'die');
+	} else {
+	    eval {
+		delete $main::str_draw{$layer};
+		delete $main::p_draw{$layer};
+		if ($type eq 'p') {
+		    $main::c->delete($layer.'-fg');
+		    $main::c->delete($layer.'-img');
+		} else {
+		    $main::c->delete($layer);
+		}
+		delete $layer_for_type_file{$type_file};
+		BBBikeLazy::bbbikelazy_remove_data($type, $layer);
+		$active = 0;
+	    };
+	    if ($@) {
+		$want_plot_type_file{$type_file} = 1;
+		main::status_message("Cannot toggle layer: $@", 'die');
+	    }
 	}
     }
     ($layer, $active);
@@ -1239,18 +1253,6 @@ sub layer_checkbutton {
 	 }
      },
     ];
-}
-
-sub add_new_nonlazy_maybe_orig_layer {
-    my($type, $file, %args) = @_;
-    add_new_nonlazy_layer($type, _maybe_orig_file($file), %args);
-}
-
-sub add_new_nonlazy_layer {
-    my($type, $file, %args) = @_;
-    require BBBikeAdvanced;
-    local $main::lazy_plot = 0; # lazy mode does not support bbd images yet
-    main::plot_additional_layer($type, $file, %args);
 }
 
 sub set_layer_highlightning {
